chore: stop sending scripts if bot detected
