--- app/styles/client.context.ts
@@ -1,13 +1,0 @@
-import { createContext } from 'react';
-
-export interface ClientStyleContextData {
-  reset: () => void;
-  sheet: string;
-}
-
-const ClientStyleContext = createContext<ClientStyleContextData>({
-  reset: () => {},
-  sheet: '',
-});
-
-export default ClientStyleContext;
